Former-commit-id: a41fa4aca2293d9c31229f599f6f254e2983dba6

+++ Sources/Internal/Scene3D/Entity.cpp
@@ -51,14 +51,14 @@
 #include "Utils/Random.h"
 #include "Scene3D/Components/ComponentHelpers.h"
 
-#define USE_VECTOR(x) (((1 << x) & vectorComponentsMask) != 0)
+#define USE_VECTOR(x) ((((uint64)1 << (uint64)x) & vectorComponentsMask) != (uint64)0)
 
 namespace DAVA
 {
 
 const int COMPONENT_COUNT_V6 = 18;
     
-uint32 vectorComponentsMask = (1 << Component::TRANSFORM_COMPONENT) | (1 << Component::RENDER_COMPONENT) | (1 << Component::LOD_COMPONENT);
+uint64 vectorComponentsMask = (1 << Component::TRANSFORM_COMPONENT) | (1 << Component::RENDER_COMPONENT) | (1 << Component::LOD_COMPONENT);
 
 // Property Names
 const char* Entity::SCENE_NODE_IS_SOLID_PROPERTY_NAME = "editor.isSolid";
